tests/wpt/tests/css/css-text/parsing/hyphenate-character-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text: getComputedStyle().hyphenateCharacter</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#hyphenate-character">
<meta name="assert" content="hyphenate-character computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("hyphenate-character", "auto");
test_computed_value("hyphenate-character", "\"=\"");
test_computed_value("hyphenate-character", "\"/-/\"");
test_computed_value("hyphenate-character", "\"\\1400\"", "\"á€\"");
test_computed_value("hyphenate-character", "\"\"");
</script>
</body>
</html>
